+++ js/utils/random.js
@@ -1,0 +1,20 @@
+const getRandomIntFromTo = (from, to) => {
+  if (from < 0 || to <= from) {
+    return false;
+  }
+  from = Math.ceil(from);
+  to = Math.floor(to);
+  return Math.floor(Math.random() * (to - from)) + from;
+};
+
+const getRandomIntFromToWithComma = (from, to, countSignsAfterComma) => {
+  if (from < 0 || to <= from) {
+    return false;
+  }
+  return (Math.random() * (to - from) + from).toFixed(countSignsAfterComma);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomIntFromTo(0, elements.length - 1)];
+
+
+export { getRandomIntFromTo, getRandomIntFromToWithComma, getRandomArrayElement };
